+++ index.html
@@ -19,6 +19,8 @@
     </style>
   </head>
   <body>
-    <div id="phaser-container"></div>
+    <div id="root">
+      <div id="phaser-container"></div>
+    </div>
   </body>
 </html>
