+++ src/Game/objects/MovingPlatform.js
