+++ src/Game/objects/Shuriken.js
