+++ src/Home/home.html
@@ -1,0 +1,6 @@
+<div>
+  메인페이지가 여기 들어간다
+</div>
+<button class="route" route="/play">
+  이걸 누르면 게임이 시작된다.
+</button>
