+++ src/Home/home.js
@@ -1,0 +1,11 @@
+import html from "./home.html";
+import "./home.css";
+
+function render() {
+  console.log("home!!");
+}
+
+export default {
+  html,
+  render,
+};
