+++ src/Play/play.html
@@ -1,0 +1,1 @@
+<div id="phaser-container"></div>
